Circle CI 2 migration
Fix intermittently failing tests
