+++ .circleci/config.yml
@@ -1,0 +1,26 @@
+version: 2
+
+jobs:
+  build:
+    docker:
+      - image: integrationcheckout/ecommerce:magento2.1
+    working_directory: ~/cko_magento2
+    steps:
+      - checkout
+      - run:
+          name: Refresh Magetno and Install Checkout
+          command: |
+            sudo a2enmod rewrite && sudo service apache2 restart && sudo service mysql restart
+            cd /var/www/html/magento2/app && sudo mkdir code && cd code && sudo mkdir CheckoutCom && cd CheckoutCom && sudo mkdir Magento2 && sudo cp -R ~/cko_magento2/* Magento2/
+            cd /var/www/html/magento2 && sudo php bin/magento setup:upgrade && sudo php bin/magento setup:di:compile && sudo chmod -R 777 .
+      - run:
+          name: Install dependencies
+          command: |
+            cd ~/cko_magento2/tests && npm i 
+      - run:
+          name: Run Tests
+          command: |
+            cd ~/cko_magento2/tests && ./node_modules/.bin/chimp config/chimp.js
+      - store_artifacts:
+          path: ~/cko_magento2/tests/screenshots
+          destination: Failed-Tests
