--- tests/config/000-default.conf
@@ -1,14 +1,0 @@
-<VirtualHost *:80>
-  ServerAdmin webmaster@localhost.com
-  DocumentRoot /var/www/html
-
-  <Directory /var/www/html>
-        Options Indexes FollowSymLinks
-        AllowOverride All
-        Require all granted
-  </Directory>
-
-  ErrorLog ${APACHE_LOG_DIR}/error.log
-  CustomLog ${APACHE_LOG_DIR}/access.log combined
-
-</VirtualHost>
