+++ tests/config/chimp.js
@@ -1,14 +1,13 @@
 
 module.exports = {
   path: './features',
-  screenshotsPath: './reports/screenshots',
+  screenshotsPath: './screenshots',
   screenshotsOnError: true,
   saveScreenshotsToDisk: true,
   saveScreenshotsToReport: true,
-  timeout: 120000,
   webdriverio: {
     logLevel: 'silent',
-    screenshotPath: './reports/screenshots',
+    screenshotPath: './screenshots',
     desiredCapabilities: {
       chromeOptions: {
         args: ['headless', 'disable-gpu'],
