+++ tests/config/globals.js
@@ -1,13 +1,12 @@
 
 export default {
-
   value: {
     url: {
       magento_base: 'http://127.0.0.1/magento2',
       admin_path: '/admin',
-      catalog_path: '/admin/catalog/product/',
+      catalog_path: '/admin/div._required:nth-child(1) > div:nth-child(2) > input:nth-child(1)/product/',
       payments_path: '/admin/admin/system_config/edit/section/payment/',
-      product_path: '/fusion-backpack.html',
+      product_path: '/ckoproduct.html',
       sign_in_path: '/customer/account/login/',
       sign_out_path: '/customer/account/logout/',
       sign_up_path: '/customer/account/create/',
@@ -15,6 +14,7 @@ export default {
       invoices_path: '/admin/sales/invoice/',
       test_product_path: '/admin/catalog/product/edit/id/6/',
       cache_path: '/admin/admin/cache/',
+      create_product_path: '/admin/catalog/product/new/set/4/type/simple/',
     },
     admin: {
       username: 'admin',
@@ -29,9 +29,9 @@ export default {
       name: 'John',
       lastname: 'Smith',
       address: '42 Ealing Broadway',
-      city: 'London',
-      country: 'GB',
-      postcode: 'W1W 8SY',
+      city: 'City',
+      county: '1',
+      postcode: '85001',
       phone: '07123456789',
     },
     customer: {
@@ -47,37 +47,37 @@ export default {
     card: {
       visa: {
         card_number: '4242424242424242',
-        mounth: '06',
+        month: '06',
         year: '18',
         cvv: '100',
       },
       mastercard: {
         card_number: '5436031030606378',
-        mounth: '06',
+        month: '06',
         year: '25',
         cvv: '257',
       },
       amex: {
         card_number: '345678901234564',
-        mounth: '06',
+        month: '06',
         year: '25',
         cvv: '1051',
       },
       diners: {
         card_number: '30123456789019',
-        mounth: '06',
+        month: '06',
         year: '25',
         cvv: '257',
       },
       jcb: {
         card_number: '3530111333300000',
-        mounth: '06',
+        month: '06',
         year: '18',
         cvv: '100',
       },
       discover: {
         card_number: '6011111111111117',
-        mounth: '06',
+        month: '06',
         year: '18',
         cvv: '100',
       },
@@ -92,6 +92,10 @@ export default {
     theme_color: '#eaeaea',
     button_label: 'johnny',
     fieldset_block: 'fieldset_block-content-loading',
+    new_product_name: 'ckoproduct',
+    new_product_price: '123',
+    new_product_quantity: '999',
+    new_product_stock: '1',
   },
   selector: {
     frontend: {
@@ -106,14 +110,13 @@ export default {
         checkout_page_loader: '#checkout-loader',
         customer_email: '#customer-email',
         email_fieldset: '#customer-email-fieldset',
-        customer_firstname: '#shipping-new-address-form > div:nth-child(1) > div:nth-child(2) > input:nth-child(1)',
-        customer_lastname: 'div._required:nth-child(2) > div:nth-child(2) > input:nth-child(1)',
-        customer_street: 'fieldset.field > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > input:nth-child(1)',
-        customer_country: 'div.field:nth-child(9) > div:nth-child(2) > select:nth-child(1)',
-        customer_uk_selector: 'div.field:nth-child(9) > div:nth-child(2) > select:nth-child(1) > option:nth-child(234)',
-        customer_city: 'div.field:nth-child(5) > div:nth-child(2) > input:nth-child(1)',
-        customer_postcode: 'div.field:nth-child(8) > div:nth-child(2) > input:nth-child(1)',
-        customer_phone: 'div.field:nth-child(10) > div:nth-child(2) > input:nth-child(1)',
+        customer_firstname: '[name="firstname"]',
+        customer_lastname: '[name="lastname"]',
+        customer_street: '[name="street[0]"]',
+        customer_county: '[name="region_id"]',
+        customer_city: '[name="city"]',
+        customer_postcode: '[name="postcode"]',
+        customer_phone: '[name="telephone"]',
         shipping_method: '#s_method_flatrate',
         terms: '#agreement_1',
         terms_hosted: 'div.checkout-agreements-block:nth-child(3) > div:nth-child(1) > div:nth-child(1) > div:nth-child(1) > input:nth-child(1)',
@@ -122,7 +125,7 @@ export default {
         checkout_iframe_selector: '#cko-iframe-id',
         embedded_fields: {
           card_number: "input[data-checkout='card-number']",
-          mounth: "input[data-checkout='expiry-month']",
+          month: "input[data-checkout='expiry-month']",
           year: "input[data-checkout='expiry-year']",
           cvv: "input[data-checkout='cvv']",
         },
@@ -152,7 +155,7 @@ export default {
       },
       hosted: {
         card_number: 'div.input-group:nth-child(2) > input:nth-child(2)',
-        mounth: 'div.split-view:nth-child(2) > div:nth-child(1) > input:nth-child(1)',
+        month: 'div.split-view:nth-child(2) > div:nth-child(1) > input:nth-child(1)',
         year: 'div.split-view:nth-child(2) > div:nth-child(2) > input:nth-child(1)',
         cvv: 'div.split-view:nth-child(3) > div:nth-child(2) > div:nth-child(1) > input:nth-child(2)',
         title: '',
@@ -240,6 +243,12 @@ export default {
       test_product_save: '#save-button',
       admin_loader: '.admin__form-loading-mask',
       flash_catch: '#flush_magento',
+      new_product_name: 'div._required:nth-child(3) > div:nth-child(2) > input:nth-child(1)',
+      new_product_price: 'div._required:nth-child(1) > div:nth-child(2) > div:nth-child(1) > input:nth-child(1)',
+      new_product_quantity: 'fieldset.admin__field:nth-child(7) > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > input:nth-child(1)',
+      new_product_stock: 'div.admin__field:nth-child(8) > div:nth-child(2) > select:nth-child(1)',
+      new_product_load_mask: '.admin__form-loading-mask',
+      save_success_message: 'div.message:nth-child(1)',
     },
   },
 };
