+++ tests/features/_setup.feature
@@ -1,0 +1,13 @@
+Feature: Prepare Magento for Tests
+      Disable SecretKey Encryption, Create a Product, Set Checkout Keys
+
+Scenario: I setup Magento for tests
+      Given I set the viewport and timeout
+      Given I disable the url secret key encryption
+      Given I create a product
+      Given I go to the backend of Checkout's plugin
+      Given I set the sandbox keys
+      Given I check the sandbox keys
+      Given I save the backend settings
+      Given I create an account
+      Then I logout from the registered customer account
