+++ tests/features/guest.feature
@@ -3,10 +3,10 @@ Feature: Guest Customer Test Suite
 
 Scenario: I should be able to complete a 3D transaction using Frames integration
       Given I go to the backend of Checkout's plugin
-      Given I enable 3D Secure
+      Given I enable THREE D
       Given I set the integration type to frames
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have frames and THREE D and no customisation
       Then I complete the order flow as a unregistered customer until the payment stage
       Then I choose Checkout as a payment option
       Then I complete Checkout Frames with a mastercard card
@@ -16,10 +16,10 @@ Scenario: I should be able to complete a 3D transaction using Frames integration
 
 Scenario: I should be able to complete a non-3D transaction using Frames integration
       Given I go to the backend of Checkout's plugin
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the integration type to frames
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have frames and no THREE D and no customisation
       Then I complete the order flow as a unregistered customer until the payment stage
       Then I choose Checkout as a payment option
       Then I complete Checkout Frames with a mastercard card
@@ -28,10 +28,10 @@ Scenario: I should be able to complete a non-3D transaction using Frames integra
 
 Scenario: I should be able to complete a 3D transaction using Hosted integration
       Given I go to the backend of Checkout's plugin
-      Given I enable 3D Secure
+      Given I enable THREE D
       Given I set the integration type to hosted
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and THREE D and no customisation
       Then I complete the order flow as a unregistered customer until the payment stage
       Then I choose Checkout as a payment option
       Then I submit the order for the hosted integration
@@ -41,10 +41,10 @@ Scenario: I should be able to complete a 3D transaction using Hosted integration
 
 Scenario: I should be able to complete a non-3D transaction using Hosted integration
       Given I go to the backend of Checkout's plugin
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the integration type to hosted
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and no customisation
       Then I complete the order flow as a unregistered customer until the payment stage
       Then I choose Checkout as a payment option
       Then I submit the order for the hosted integration
