+++ tests/features/registered.feature
@@ -3,10 +3,10 @@ Feature: Registered Customer Test Suite
 
 Scenario: I should be able to complete a 3D transaction using Frames integration
       Given I go to the backend of Checkout's plugin
-      Given I enable 3D Secure
+      Given I enable THREE D
       Given I set the integration type to frames
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have frames and THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
@@ -18,10 +18,10 @@ Scenario: I should be able to complete a 3D transaction using Frames integration
 
 Scenario: I should be able to complete a non-3D transaction using Frames integration
       Given I go to the backend of Checkout's plugin
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the integration type to frames
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have frames and no THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
@@ -32,10 +32,10 @@ Scenario: I should be able to complete a non-3D transaction using Frames integra
 
 Scenario: I should be able to complete a 3D transaction using Hosted integration
       Given I go to the backend of Checkout's plugin
+      Given I enable THREE D
       Given I set the integration type to hosted
-      Given I enable 3D Secure
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
@@ -47,10 +47,10 @@ Scenario: I should be able to complete a 3D transaction using Hosted integration
 
 Scenario: I should be able to complete a non-3D transaction using Hosted integration
       Given I go to the backend of Checkout's plugin
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the integration type to hosted
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
@@ -61,24 +61,26 @@ Scenario: I should be able to complete a non-3D transaction using Hosted integra
 
 Scenario: I should be able to customise Hosted integration
       Given I go to the backend of Checkout's plugin
+      Given I disable THREE D
       Given I set the integration type to hosted
       Given I set the theme color
       Given I set the button label
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
       Then I submit the order for the hosted integration
       Then I should see a customised hosted page
 
-
 Scenario: I should be able to set the Checkout plugin paymet option title as the first payment option
       Given I go to the backend of Checkout's plugin
+      Given I disable THREE D
+      Given I set the integration type to hosted
       Given I set the payment option title
       Given I set the payment option order
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I should see the plugin title
@@ -86,10 +88,10 @@ Scenario: I should be able to set the Checkout plugin paymet option title as the
 Scenario: In Checkout Hosted I should be able see alternaive payments for the mixed payment mode
       Given I go to the backend of Checkout's plugin
       Given I set the integration type to hosted
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the payment mode to mixed
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
@@ -100,10 +102,10 @@ Scenario: In Checkout Hosted I should be able see alternaive payments for the mi
 Scenario: In Checkout Hosted I should be able see only alternaive payments for the coresponding payment mode
       Given I go to the backend of Checkout's plugin
       Given I set the integration type to hosted
-      Given I disable 3D Secure
+      Given I disable THREE D
       Given I set the payment mode to cards
       Given I save the backend settings
-      Then I clear magento's cache
+      Given I have hosted and no THREE D and no customisation
       Given I login the registered customer account
       Then I complete the order flow as a registered customer until the payment stage
       Then I choose Checkout as a payment option
