+++ tests/package.json
@@ -1,22 +1,16 @@
 {
-  "name": "magento-2.1",
+  "name": "e2e_magento2",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "lint": "./node_modules/.bin/eslint **/**.js",
-    "test": "./node_modules/.bin/chimp ./config/chimp.js",
-    "watch": "./node_modules/.bin/chimp ./config/chimp.js --watch"
+    "test": "node --max-old-space-size=2000 ./node_modules/.bin/chimp ./config/chimp.js --watch"
   },
   "keywords": [],
-  "author": "“Johnny <“ionutghisoi95@gmail.com”>",
+  "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "chimp": "^0.50.2",
-    "eslint": "^4.11.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.5.1"
+  "dependencies": {
+    "chai": "^4.1.2",
+    "chimp": "^0.50.2"
   }
 }
